refactor: cli and client.rs
