+++ src-tauri/src/bin/rssh/commands/mod.rs
@@ -1,0 +1,7 @@
+pub mod add;
+pub mod completions;
+pub mod config;
+pub mod edit;
+pub mod ls;
+pub mod open;
+pub mod rm;
