+++ src-tauri/src/lib.rs
@@ -43,11 +43,16 @@ pub fn run() {
 
             // 一次性清理：旧版曾把私钥 passphrase 落盘到 SecretStore。
             // 新流程改为终端内交互输入 + 进程内缓存，旧条目永不读，全部删掉。
+            // 用 setting 标志记录已迁移，避免每次启动都遍历 keychain 删 N 次。
             // 失败不致命：删不掉的项最坏情况是占几个字节，不影响功能。
-            if let Ok(creds) = db::credential::list(&db) {
-                for c in creds {
-                    let _ = secret_store.delete(&secret::cred_passphrase_key(&c.id));
+            const MIGRATION_KEY: &str = "migration_passphrase_keychain_cleared_v1";
+            if matches!(db::settings::get(&db, MIGRATION_KEY), Ok(None)) {
+                if let Ok(creds) = db::credential::list(&db) {
+                    for c in creds {
+                        let _ = secret_store.delete(&secret::cred_passphrase_key(&c.id));
+                    }
                 }
+                let _ = db::settings::set(&db, MIGRATION_KEY, "1");
             }
 
             app.manage(AppState {
@@ -115,6 +120,7 @@ pub fn run() {
             commands::session::ssh_resize,
             commands::session::ssh_disconnect,
             commands::session::ssh_auth_respond,
+            commands::session::ssh_auth_cancel,
             commands::session::ssh_passphrase_respond,
             commands::session::ssh_passphrase_cancel,
             commands::session::ssh_host_key_respond,
