+++ src-tauri/src/ssh/mod.rs
@@ -1,6 +1,8 @@
+pub mod auth;
 pub mod bastion;
 pub mod client;
 pub mod config;
 pub mod forward;
 pub mod known_hosts;
+pub mod prompt;
 pub mod sftp;
